shared/shared/middleware/request_id.py
import uuid
from collections.abc import Awaitable, Callable

from starlette.requests import Request
from starlette.responses import Response


async def request_id_middleware(
    request: Request, call_next: Callable[[Request], Awaitable[Response]]
) -> Response:
    request_id = request.headers.get("X-Request-ID") or str(uuid.uuid4())
    request.state.request_id = request_id
    response = await call_next(request)
    response.headers["X-Request-ID"] = request_id
    return response
