shared/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docfliq-shared"
version = "0.1.0"
description = "Shared auth, events, models, database, middleware for Docfliq services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-jose[cryptography]>=3.3",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-asyncio>=0.23", "ruff>=0.1"]

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]
